Moved classes, used LinkedList size constant
