+++ jre_emul/android/platform/libcore/ojluni/src/objc/java/java/util/Collections.java
